pyproject.toml
[project]
name = "job-application-automation"
version = "1.0.0"
description = "AutoApply AI: Smart Job Application Assistant - Automates job search, resume generation, and application submission"
requires-python = ">=3.10"
dependencies = [
    # Core
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
    "python-json-logger>=2.0.7",
    "typing-extensions>=4.7.0",

    # Database & ORM
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # Browser automation
    "browser-use>=0.1.40",
    "playwright>=1.40.0",
    "selenium>=4.15.0",
    "undetected-chromedriver>=3.5.0",

    # Web scraping
    "crawl4ai>=0.5.0.post8",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "requests>=2.31.0",
    "httpx>=0.24.1",
    "aiohttp>=3.8.5",

    # Document processing
    "python-docx>=1.0.0",
    "docxtpl>=0.16.7",
    "markdown>=3.4.0",
    "PyPDF2>=3.0.0",
    "pdf2image>=1.16.0",

    # AI/LLM integration
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
    "azure-ai-inference>=1.0.0b7",

    # ML/NLP
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "spacy>=3.7.0",
    "fuzzywuzzy>=0.18.0",
    "textblob>=0.18.0",

    # Data & Reporting
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",

    # Utility
    "tqdm>=4.66.0",
    "cachetools>=5.3.0",
    "click>=8.1.0",
    "colorama>=0.4.6",
    "rich>=13.7.0",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",

    # API server
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.36.0",

    # Monitoring
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "mysqlclient>=2.2.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["job_application_automation*"]

[tool.pytest.ini_options]
testpaths = ["job_application_automation/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
